% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaHjelpefunk.R
\name{abonnementKorona}
\alias{abonnementKorona}
\title{Funksjon som produserer rapporten som skal sendes til mottager.
(The actual call to this function is made through do.call and
has the effect of providing the parameters as class
\emph{list}. Verdier gis inn som listeparametre}
\usage{
abonnementKorona(
  rnwFil,
  brukernavn = "lluring",
  reshID = 0,
  Rpakke = "korona",
  valgtEnhet = "Alle",
  enhetsNivaa = "RHF",
  rolle = "SC"
)
}
\arguments{
\item{rnwFil}{Navn på fila som skal kjøres. Angis MED filending (\emph{dvs "filnavn.Rnw"})}

\item{reshID}{Aktuell reshid}

\item{Rpakke}{hvilken R-pakke fila som lager rapporten ligger i}

\item{datoFra}{dato}

\item{parametre}{Liste med valgfrie parametre, avhengig av type rapport}
}
\value{
Full path of file produced
}
\description{
Funksjon som produserer rapporten som skal sendes til mottager.
(The actual call to this function is made through do.call and
has the effect of providing the parameters as class
\emph{list}. Verdier gis inn som listeparametre
}
