% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaFigurer.R
\name{AlderKjFig}
\alias{AlderKjFig}
\title{Aldersfordeling, tabell}
\usage{
AlderKjFig(
  RegData,
  valgtVar = "Alder",
  valgtEnhet = "Alle",
  enhetsNivaa = "RHF",
  datoFra = "2020-03-01",
  datoTil = Sys.Date(),
  skjemastatusInn = 9,
  minN = 0,
  aarsakInn = 9,
  dodSh = 9,
  erMann = 9,
  grvar = "erMann",
  outfile = ""
)
}
\arguments{
\item{RegData}{datatabell, beredskapsdata}

\item{valgtEnhet}{- gis ut fra resh}

\item{enhetsNivaa}{organisatorisk nivå, dvs. filtreringsnivå for egen enhet.
RHF (SC og LC-brukere), HF (LU-brukere)}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{skjemastatusInn}{status på inklusjonsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{minN}{minste antall: Maskerer verdier under minN, standard: 0 (ingen maskering)}

\item{aarsakInn}{covid-19 som hovedårsak til innleggelse 1-ja, alle opph, 2-ja, minst siste opphold,
3-ja, minst ett opph, 4-nei, ingen opph, 9-ukj}

\item{dodSh}{død på sykehus 1-'nei', 2-'ja', 3-'ja og nei' (=ferdigstilte?)}

\item{erMann}{kjønn, 0-kvinne, 1-mann}
}
\value{
Alders- og kjønnsfordeling
}
\description{
Aldersfordeling, tabell
}
