% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{AlderTab}
\alias{AlderTab}
\title{Aldersfordeling, tabell}
\usage{
AlderTab(
  RegData,
  valgtEnhet = "Alle",
  enhetsNivaa = "RHF",
  skjemastatusInn = 9,
  aarsakInn = 9,
  dodSh = 9,
  erMann = 9
)
}
\arguments{
\item{RegData}{datatabell, beredskapsdata}

\item{valgtEnhet}{- gis ut fra resh}

\item{enhetsNivaa}{styres av tilgangsnivå 'Alle', 'RHF', 'HF'}

\item{skjemastatusInn}{status på inklusjonsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{aarsakInn}{covid-19 som hovedårsak til innleggelse 1-ja, 2-nei}

\item{dodSh}{død på sykehus 1-'nei', 2-'ja', 3-'ja og nei' (=ferdigstilte?)}

\item{erMann}{kjønn, 0-kvinne, 1-mann}
}
\value{

}
\description{
Aldersfordeling, tabell
}
