% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{antallTidEnhTab}
\alias{antallTidEnhTab}
\title{Antall tilfeller for valgt tidsenhet og enhetsnivå. Filtreringer kan også gjøres.
Detaljerinsnivå er styrt av tilgangsnivå}
\usage{
antallTidEnhTab(
  RegData,
  tidsenhet = "dag",
  erMann = 9,
  tilgangsNivaa = "SC",
  skjemastatusInn = 9,
  aarsakInn = 9,
  dodSh = 9,
  valgtEnhet = "Alle"
)
}
\arguments{
\item{RegData}{dataramme med preprossesserte data}

\item{tidsenhet}{'dag' (standard), 'uke', 'maaned'}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{tilgangsNivaa}{SC, LC og LU bestemmer hvilket enhetsNivaa
('RHF', 'HF', 'ShNavn') resultatene skal vises for}

\item{skjemastatusInn}{status på inklusjonsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{aarsakInn}{covid-19 som hovedårsak til innleggelse}

\item{dodSh}{død på sykehus 0-nei, 1-ja}

\item{valgtEnhet}{NULL for SC-bruker, ellers eget RHF/HF}
}
\value{

}
\description{
Antall tilfeller for valgt tidsenhet og enhetsnivå. Filtreringer kan også gjøres.
Detaljerinsnivå er styrt av tilgangsnivå
}
