% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabellerK2n.R
\name{antallTidInneliggende}
\alias{antallTidInneliggende}
\title{Antall inneliggende per tidsenhet  og enhetsnivå. Filtreringer kan også gjøres.
Detaljerinsnivå er styrt av tilgangsnivå}
\usage{
antallTidInneliggende(
  RegData,
  tidsenhet = "dag",
  erMann = 9,
  tilgangsNivaa = "SC",
  datoFra = 0,
  datoTil = Sys.Date(),
  skjemastatusInn = 9,
  aarsakInn = 9,
  valgtEnhet = "Alle"
)
}
\arguments{
\item{RegData}{dataramme med preprossesserte data, NB: IKKE personaggregert}

\item{tidsenhet}{'dag' (standard), 'uke', 'maaned'}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{tilgangsNivaa}{SC, LC og LU bestemmer hvilket enhetsNivaa
('RHF', 'HF', 'ShNavn') resultatene skal vises for}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{skjemastatusInn}{status på inklusjonsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{aarsakInn}{1-ja, 2-nei, 9-alle reg}

\item{valgtEnhet}{NULL for SC-bruker, ellers eget RHF/HF}
}
\value{

}
\description{
Antall inneliggende per tidsenhet  og enhetsnivå. Filtreringer kan også gjøres.
Detaljerinsnivå er styrt av tilgangsnivå
}
