% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabellerK2n.R
\name{antallTidUtskrevne}
\alias{antallTidUtskrevne}
\title{Antall utskrevne per tidsenhet (basert på avdøddato) og enhetsnivå. Filtreringer kan også gjøres.
Detaljerinsnivå er styrt av tilgangsnivå}
\usage{
antallTidUtskrevne(
  RegData,
  tidsenhet = "dag",
  erMann = 9,
  tilgangsNivaa = "SC",
  datoFra = 0,
  skjemastatusInn = 9,
  aarsakInn = 9,
  valgtEnhet = "Alle"
)
}
\arguments{
\item{RegData}{dataramme med preprossesserte data}

\item{tidsenhet}{'dag' (standard), 'uke', 'maaned'}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{tilgangsNivaa}{SC, LC og LU bestemmer hvilket enhetsNivaa
('RHF', 'HF', 'ShNavn') resultatene skal vises for}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{skjemastatusInn}{status på inklusjonsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{aarsakInn}{covid-19 som hovedårsak til innleggelse 1-ja, 2-nei}

\item{valgtEnhet}{NULL for SC-bruker, ellers eget RHF/HF}
}
\value{

}
\description{
Antall utskrevne per tidsenhet (basert på avdøddato) og enhetsnivå. Filtreringer kan også gjøres.
Detaljerinsnivå er styrt av tilgangsnivå
}
