% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabellerK2n.R
\name{erInneliggende}
\alias{erInneliggende}
\title{Funksjon som avgjør om en pasient er inneliggende på aktuell dato}
\usage{
erInneliggende(aktuelldato)
}
\arguments{
\item{aktuelldato}{dato som inneligging skal avgjøres for}

\item{InnDato}{Innleggelsesdato}

\item{UtDato}{Utskrivingsdato}
}
\value{

}
\description{
Returnerer TRUE for datoer pasienten er inneliggende
}
