% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{FerdigeRegInnTab}
\alias{FerdigeRegInnTab}
\title{Ferdigstilte registreringer}
\usage{
FerdigeRegInnTab(
  RegData,
  valgtEnhet = "Alle",
  enhetsNivaa = "RHF",
  erMann = 9,
  dodSh = 9
)
}
\arguments{
\item{RegData}{beredskapsskjema}

\item{valgtEnhet}{- gis ut fra resh}

\item{enhetsNivaa}{organisatorisk nivå, dvs. filtreringsnivå for egen enhet.
RHF (SC og LC-brukere), HF (LU-brukere)}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{dodSh}{død på sykehus 1-'nei', 2-'ja', 3-'ja og nei' (=ferdigstilte?)}
}
\value{

}
\description{
Ferdigstilte registreringer
}
