% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{FerdigeRegTab}
\alias{FerdigeRegTab}
\title{Ferdigstilte registreringer, utskrevne pasienter}
\usage{
FerdigeRegTab(
  RegData,
  valgtEnhet = "Alle",
  enhetsNivaa = "RHF",
  minN = 0,
  datoFra = "2020-03-01",
  datoTil = Sys.Date(),
  aarsakInn = 9,
  erMann = 9,
  dodSh = 9
)
}
\arguments{
\item{RegData}{korona-registreringer}

\item{valgtEnhet}{- gis ut fra resh}

\item{enhetsNivaa}{organisatorisk nivå, dvs. filtreringsnivå for egen enhet.
RHF (SC og LC-brukere), HF (LU-brukere)}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{aarsakInn}{covid-19 som hovedårsak til innleggelse 1-ja, alle opph, 2-ja, minst siste opphold,
3-ja, minst ett opph, 4-nei, ingen opph, 9-ukj}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{dodSh}{død på sykehus 1-'nei', 2-'ja', 3-'ja og nei' (=ferdigstilte?)}
}
\description{
Ferdigstilte registreringer, utskrevne pasienter
}
