% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaFigurer.R
\name{FigTidEnhet}
\alias{FigTidEnhet}
\title{Antall tilfeller for valgt tidsenhet og enhetsnivå. Filtreringer kan også gjøres.}
\usage{
FigTidEnhet(AntTab, outfile = "")
}
\arguments{
\item{AntTab}{Dataramme med nødvendige figurparametre}
}
\value{

}
\description{
Antall tilfeller for valgt tidsenhet og enhetsnivå. Filtreringer kan også gjøres.
}
