% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{finnBeredUpandemi}
\alias{finnBeredUpandemi}
\title{Tabell med oversikt over hvilke beredskapsskjema som mangler pandemiskjema}
\usage{
finnBeredUpandemi(
  datoFra = "2020-01-01",
  datoTil = Sys.Date(),
  HF = "Alle",
  ...
)
}
\arguments{
\item{datoFra}{- startdato for innleggelse på intensiv}
}
\value{
Beredskapsskjema som mangler pandemiskjema
}
\description{
Tabell med oversikt over hvilke beredskapsskjema som mangler pandemiskjema
}
