% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunksjonerDatafilerFHI.R
\name{hentBeredDataFHI}
\alias{hentBeredDataFHI}
\title{Henter data og velger variabler for overføring til FHI}
\usage{
hentBeredDataFHI(personIDvar = "PersonIdBC19Hash")
}
\value{
Beredskapsdata tilrettelagt for FHI
}
\description{
Henter data og velger variabler for overføring til FHI
}
