% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaHjelpefunk.R
\name{henteSamlerapporterKorona}
\alias{henteSamlerapporterKorona}
\title{Funksjon som produserer rapporten som skal lastes ned av mottager.}
\usage{
henteSamlerapporterKorona(
  filnavn,
  rnwFil,
  Rpakke = "korona",
  rolle = "SC",
  valgtEnhet = "Alle",
  enhetsNivaa = "RHF",
  reshID = 0
)
}
\arguments{
\item{filnavn}{dummy}

\item{rnwFil}{Navn på fila som skal kjøres. Angis uten ending (\emph{dvs uten  ".Rnw"})}

\item{Rpakke}{hvilken R-pakke fila som lager rapporten ligger i}

\item{datoFra}{dato}
}
\value{
Full path of file produced
}
\description{
Funksjon som produserer rapporten som skal lastes ned av mottager.
}
