% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunksjonerDatafilerFHI.R
\name{hentPandemiDataFHI}
\alias{hentPandemiDataFHI}
\title{Henter data og velger variabler for overføring til FHI}
\usage{
hentPandemiDataFHI(personIDvar = "PersonIdBC19Hash", raa = 1, aggP = 1)
}
\arguments{
\item{personIDvar}{angi variabel for personid
'PersonIdBC19Hash' - standard, 'PersonId' - alternativ}

\item{raa}{hente rådata= 1-ja (standard), 0-nei}

\item{aggP}{hente persondata aggregert til smitteforløp? 1-ja (standard), 0-nei}
}
\value{
Pandemidata tilrettelagt for FHI
}
\description{
Henter data og velger variabler for overføring til FHI
}
