% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{innManglerUt}
\alias{innManglerUt}
\title{Finn innleggelser som mangler utskriving}
\usage{
innManglerUt(RegData, valgtEnhet = "Alle", enhetsNivaa = "RHF")
}
\arguments{
\item{RegData}{Koblet dataramme uten aggregering}

\item{valgtEnhet}{egen/valgt enhet}

\item{enhetsNivaa}{eget/valgt enhetsnivå}
}
\description{
Finn innleggelser som mangler utskriving
}
