% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaDataSQL.R
\name{KoronaDataSQL}
\alias{KoronaDataSQL}
\title{SKAL Hente data fra database for koronaregisteringer.NB MÅ OPPDATERES}
\usage{
KoronaDataSQL(skjema = 1)
}
\arguments{
\item{skjema}{1-innleggelse, 2-utskriving}
}
\value{
Henter dataramma RegData for Intensivregisteret
}
\description{
SKAL Hente data fra database for koronaregisteringer.NB MÅ OPPDATERES
}
