% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaFigAndeler.R
\name{KoronaFigAndeler}
\alias{KoronaFigAndeler}
\title{Funksjon som beregner aggregerte verdier (andeler) for ulike variabler/variabelkombinasjoner}
\usage{
KoronaFigAndeler(
  RegData = 0,
  valgtVar = "alder",
  datoFra = "2011-01-01",
  datoTil = "3000-12-31",
  aar = 0,
  overfPas = 0,
  minald = 0,
  maxald = 110,
  erMann = "",
  InnMaate = "",
  dodInt = "",
  outfile = "",
  grType = 99,
  preprosess = 1,
  hentData = 0,
  reshID = 0,
  velgAvd = 0,
  enhetsUtvalg = 0,
  lagFig = 1,
  ...
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{overfPas}{Overført under pågående intensivbehandling? 
1 = Pasienten er ikke overført
2 = Pasienten er overført}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 110)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{dodInt}{Levende/død ut fra intensiv. 
0: i live, 
1: død,   
alle: standard (alle andre verdier)}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{lagFig}{Angir om figur skal lages eller ikke 0-ikke lag, 1-lag}

\item{figurtype}{Hvilken figurtype som ønskes ut: 
andel (fordelingsfigurer), 
andelGrVar (andel i hver kategori av grupperingsvariabel, eks. sykehus), 
andelTid (andel per tidsenhet, eks. år, måned), 
andelPP (andel før og etter), 
gjsnGrVar (sentralmål i hver kategori av grupperingsvariabel, eks. sykehus),
gjsnTid (sentralmål per tidsenhet, eks. år, måned)}
}
\value{
Søylediagram (fordeling) av valgt variabel. De enkelte verdiene kan også sendes med.
}
\description{
Denne funksjonen beregner AggVerdier (fordeling) av valgt variabel
filtrert på de utvalg som er gjort. Kan trenge funksjonerne:
}
\details{
Funksjonen benytter funksjonene: NIRRegDataSQL, NIRPreprosess, NIRVarTilrettelegg, NIRUtvalgEnh
og NIRFigSoyler

Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Aldersfordeling, 10-årige grupper 
    \item inklKrit: Andeler for de 5 inklusjonskriteriene
    \item liggetid: Liggetid 
    \item PrimaryReasonAdmitted: Hovedårsak til intensivopphold
    \item respiratortid: Tid tilbrakt i respirator
    \item spesTiltak: Spesielle tiltak
   }
}
