% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaFigAndelTid.R
\name{KoronaFigAndelTid}
\alias{KoronaFigAndelTid}
\title{Tidstrend av andel opphold}
\usage{
KoronaFigAndelTid(
  RegData = 0,
  hentData = 0,
  valgtVar = "alder_u18",
  datoFra = "2020-03-01",
  datoTil = Sys.Date(),
  tidsenhet = "Kvartal",
  dod = "",
  reshID = 0,
  erMann = 9,
  minald = 0,
  maxald = 110,
  skjemastatusInn = 9,
  skjemastatusUt = 9,
  dodSh = 9,
  aarsakInn = 9,
  enhetsNivaa = "RHF",
  valgtEnhet = "Alle",
  enhetsUtvalg = 0,
  beredPas = 9,
  outfile = "",
  lagFig = 1,
  ...
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.
' @param outfile Navn på fil figuren skrives til. Standard: '' (Figur skrives
   til systemets standard utdataenhet (som regel skjerm))}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{tidsenhet}{Oppløsning på tidsaksen. Verdier: 'Aar' (standard), 'Halvaar', 'Kvartal','Mnd'}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{minald}{kjønn, 0-kvinne, 1-mann}

\item{skjemastatusInn}{status på inklusjonsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{skjemastatusUt}{status på utskrivingsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{dodSh}{død på sykehus 1-'nei', 2-'ja', 3-'ja og nei' (=ferdigstilte?)}

\item{aarsakInn}{covid-19 som hovedårsak til innleggelse 1-ja, alle opph, 2-ja, minst siste opphold,
3-ja, minst ett opph, 4-nei, ingen opph, 9-ukj}

\item{enhetsNivaa}{organisatorisk nivå, dvs. filtreringsnivå for egen enhet.
RHF (SC og LC-brukere), HF (LU-brukere)}

\item{valgtEnhet}{- gis ut fra resh}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{lagFig}{Angir om figur skal lages eller ikke 0-ikke lag, 1-lag}
}
\value{
Figur som viser tidstrend, dvs. andel av valgt variabel for hvert år.
}
\description{
Denne funksjonen lager et linjediagram som viser utvikling over tid  for andeler av valgt variabel,
filtrert på de utvalg som er gjort.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder_u18: Pasienter under 18 år
    \item alder_over80: Pasienter over 80 år (>=80)
    \item dod30d: Pasienter som dør innen 30 dager etter innleggelse
 \item isolering: Isolasjon av pasient
   }
}
