% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaFigGjsnTid.R
\name{KoronaFigGjsnTid}
\alias{KoronaFigGjsnTid}
\title{Utvikling over tid for gjennomsnitt/median av valgt variabel}
\usage{
KoronaFigGjsnTid(
  RegData = 0,
  hentData = 0,
  preprosess = 1,
  valgtVar = "alder",
  datoFra = "2020-03-01",
  datoTil = Sys.Date(),
  tidsenhet = "Kvartal",
  dod = "",
  reshID = 0,
  erMann = 9,
  minald = 0,
  maxald = 110,
  skjemastatusInn = 9,
  skjemastatusUt = 9,
  dodSh = 9,
  aarsakInn = 9,
  enhetsNivaa = "RHF",
  valgtEnhet = "Alle",
  enhetsUtvalg = 0,
  beredPas = 9,
  valgtMaal = "Gjsn",
  outfile = "",
  lagFig = 1,
  ...
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{valgtVar}{Hvilken variabel som skal visualiseres}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{minald}{kjønn, 0-kvinne, 1-mann}

\item{skjemastatusInn}{status på inklusjonsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{skjemastatusUt}{status på utskrivingsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{dodSh}{død på sykehus 1-'nei', 2-'ja', 3-'ja og nei' (=ferdigstilte?)}

\item{aarsakInn}{covid-19 som hovedårsak til innleggelse 1-ja, alle opph, 2-ja, minst siste opphold,
3-ja, minst ett opph, 4-nei, ingen opph, 9-ukj}

\item{enhetsNivaa}{organisatorisk nivå, dvs. filtreringsnivå for egen enhet.
RHF (SC og LC-brukere), HF (LU-brukere)}

\item{valgtEnhet}{- gis ut fra resh}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{beredPas}{Har pasienten vært innlagt på intensiv? 0-nei, 1-ja.
Per nå (12.03.23) basert på at faktisk finner et tilhørende beredskapsskjema.Vurder om
vi heller skal basere på variabelen Nir_beredskapsskjema_CoV2}

\item{valgtMaal}{'gjsn': gir middelverdi (standard)
'med': gir median}

\item{lagFig}{Angir om figur skal lages eller ikke 0-ikke lag, 1-lag}
}
\value{
Linjediagram som viser utvikling over tid for valgt variabel
}
\description{
Figuren viser gjennomsnitt/median per år med konfidensintervall for valgt variabel.
I bakgrunn vises konfidensintervall for det man har valt å sammenlikne med.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Pasientens alders
   }
}
