% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaPreprossesser.R
\name{KoronaPreprosesser}
\alias{KoronaPreprosesser}
\title{Preprosesser data fra Koronaregisteret}
\usage{
KoronaPreprosesser(RegData = RegData)
}
\arguments{
\item{RegData}{Koronaskjema}

\item{skjema}{1: innleggelse, 2: utskriving,}
}
\value{
Preprosesserte data
}
\description{
Denne funksjonen navner om variabler og beregner evt. nye.
}
