% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaUtvalg.R
\name{KoronaUtvalg}
\alias{KoronaUtvalg}
\title{Gjøre utvalg i koronaskjema}
\usage{
KoronaUtvalg(
  RegData,
  datoFra = 0,
  datoTil = 0,
  erMann = 9,
  minald = 0,
  maxald = 110,
  bekr = 9,
  skjemastatus = 9,
  dodInt = 9,
  reshID = 0,
  valgtRHF = "Alle"
)
}
\arguments{
\item{RegData}{data, beredskapsskjema}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{bekr}{1-bekreftet, 0-mistenkt tilfelle}

\item{skjemastatus}{status på registreringa 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{dodInt}{død på intensiv 0-nei, 1-ja}

\item{reshID}{reshID fra innlogging}

\item{enhetsUtvalg}{enhetsutvalg...}
}
\value{

}
\description{
Returnerer filtrert dataramme og utvalgstekst.
Aktuelt med utvalg på HF og RHF-nivå
}
