% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatafilerFHI_slettes.R
\name{lagDatafilerTilFHIgml}
\alias{lagDatafilerTilFHIgml}
\title{Henter data og tilrettelegger filer for overføring til FHI}
\usage{
lagDatafilerTilFHIgml()
}
\value{
datafiler
}
\description{
Henter data og tilrettelegger filer for overføring til FHI
}
