% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunksjonerDatafilerFHI.R
\name{lagInfluDataFHI}
\alias{lagInfluDataFHI}
\title{Henter data og velger variabler for overføring til FHI}
\usage{
lagInfluDataFHI(personIDvar = "PersonIdBC19Hash")
}
\value{
#Influensadata tilrettelagt for FHI
}
\description{
Henter data og velger variabler for overføring til FHI
}
