% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluensadataFHI.R
\name{lagInfluDataFHI_flyttet}
\alias{lagInfluDataFHI_flyttet}
\title{Henter data og velger variabler for overføring til FHI}
\usage{
lagInfluDataFHI_flyttet()
}
\value{
data
}
\description{
Henter data og velger variabler for overføring til FHI
}
