% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{lagTabavFigFord}
\alias{lagTabavFigFord}
\title{Vise figurdata som tabell}
\usage{
lagTabavFigFord(UtDataFraFig)
}
\arguments{
\item{UtDataFraFig}{data fra figurfunksjoner, dvs. beregnede verdier}
}
\description{
Vise figurdata som tabell
}
