% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaHjelpefunk.R
\name{LeggTilNyInnOverf}
\alias{LeggTilNyInnOverf}
\title{Beregner og legger til variabel for reinnleggelse, ny innleggelse og overføring
NB: Basert på opphold, ikke aggregerte tall!}
\usage{
LeggTilNyInnOverf(RegData, PasientID = "PasientID")
}
\arguments{
\item{RegData}{Dataramme. RegData må inneholde Inntidspunkt, Uttidspunkt og PasientID}

\item{PasientID}{variabelnavn for pasient-id}
}
\value{
Dataramme som inneholder de beregnede variablene
}
\description{
Reinn: ny innleggelse, def. >= 24t, uansett enhet
Overf: overført, def. < 24t ikke samme enhet
}
