% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{oppsumFerdigeRegTab}
\alias{oppsumFerdigeRegTab}
\title{Ferdigstilte registreringer}
\usage{
oppsumFerdigeRegTab(
  RegData,
  valgtRHF = "Alle",
  bekr = 9,
  erMann = 9,
  dodInt = 9
)
}
\arguments{
\item{RegData}{beredskapsskjema}

\item{bekr}{1-bekreftet, 0-mistenkt tilfelle}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{dodInt}{død på intensiv 0-nei, 1-ja}
}
\value{

}
\description{
Ferdigstilte registreringer
}
