% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{PasMdblReg}
\alias{PasMdblReg}
\title{Finner pasienter med dobbeltregistrerte skjema}
\usage{
PasMdblReg(RegData, tidsavvik = 0)
}
\arguments{
\item{RegData}{dataramme fra pandemi registeret, inn og utskr.skjema}

\item{tidssavik}{- maks tidsavvik (minutter) mellom to påfølgende registreringer som sjekkes}
}
\value{
dobbeltregistrering av inn-skjema
}
\description{
Finner pasienter med dobbeltregistrerte skjema
}
