% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{RisikoInnTab}
\alias{RisikoInnTab}
\title{Tabell med oversikt over tilstander som medfører økt risiko ved Coronasmitte}
\usage{
RisikoInnTab(
  RegData,
  tidsenhet = "Totalt",
  datoTil = Sys.Date(),
  reshID = 0,
  erMann = "",
  bekr = 9,
  skjemastatus = 9,
  dodSh = 9,
  valgtEnhet = "Alle",
  enhetsNivaa = "RHF",
  minald = 0,
  maxald = 110,
  velgAvd = 0
)
}
\arguments{
\item{RegData}{data}

\item{tidsenhet}{'Dag', 'Uke' (standard)}

\item{datoTil}{sluttdato}

\item{reshID}{enhetens resh}

\item{valgtEnhet}{'Alle' (standard), RHF-navn uten 'Helse '}
}
\value{

}
\description{
Tabell med oversikt over tilstander som medfører økt risiko ved Coronasmitte
}
