% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunksjonerDatafilerFHI.R
\name{sendDataFilerFHI}
\alias{sendDataFilerFHI}
\title{Funksjon som henter filer som skal sendes til FHI. To filer fra intensivopphold
og to filer fra sykehusopphold. Dvs. Ei fil for hvert opphold og ei aggregert til
person, for hvert register}
\usage{
sendDataFilerFHI(
  zipFilNavn = "Testfil",
  brukernavn = "testperson",
  recipient = "nhn"
)
}
\arguments{
\item{zipFilNavn}{Navn på fila som skal kjøres. DataFHICovMonitor, DataFHIPanBeredInflu, Testfil}

\item{brukernavn}{Innlogget brukernavn}

\item{recipient}{Character string: brukernavn for unik definisjon av mottager. Benyttes i sship.
recipient er også hardkodet ut fra hvilken filpakke som er valgt, men må kunne velges for å sende testfil til valgt mottager.
Standard: 'nhn' Valg: 'nhn', 'nhn_covmonitor'}
}
\value{
Filsti til fil med filsti til zip...
}
\description{
Funksjon som henter filer som skal sendes til FHI. To filer fra intensivopphold
og to filer fra sykehusopphold. Dvs. Ei fil for hvert opphold og ei aggregert til
person, for hvert register
}
