% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaHjelpefunk.R
\name{SorterOgNavngiTidsEnhet}
\alias{SorterOgNavngiTidsEnhet}
\title{Tilrettelegge tidsenhetvariabel:}
\usage{
SorterOgNavngiTidsEnhet(RegData, tidsenhet = "Mnd")
}
\arguments{
\item{RegData}{dataramme}

\item{tidsenhet}{tidsenhet: 'Mnd' (standard), 'Kvartal', 'Halvaar', 'Aar',}
}
\description{
Tilrettelegge tidsenhetvariabel:
}
