% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{statusNaaTab}
\alias{statusNaaTab}
\title{Nøkkeltall}
\usage{
statusNaaTab(RegData, valgtEnhet = "Alle", enhetsNivaa = "RHF", erMann = 9)
}
\arguments{
\item{RegData}{pandemiskjema}
}
\value{

}
\description{
Nøkkeltall
}
