% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{TabAlder}
\alias{TabAlder}
\title{Aldersfordeling, tabell}
\usage{
TabAlder(
  RegData,
  valgtRHF = "Alle",
  bekr = 9,
  skjemastatus = 9,
  dodInt = 9,
  erMann = 9
)
}
\arguments{
\item{RegData}{datatabell, beredskapsdata}

\item{bekr}{1-bekreftet, 0-mistenkt tilfelle}

\item{skjemastatus}{status på registreringa 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{dodInt}{død på intensiv 0-nei, 1-ja}

\item{erMann}{kjønn, 0-kvinne, 1-mann}
}
\value{

}
\description{
Aldersfordeling, tabell
}
\examples{
TabAlder(RegData=CoroData, enhetsNivaa='HF')
}
