% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{tabAntOpphEnhTid}
\alias{tabAntOpphEnhTid}
\title{tabAntOpphEnhTid antall opphold siste X (antMnd) mnd
RegData må inneholde ikke-aggregerte data, dvs. data på oppholdsnivå
Summerer antall opphold for hele tidsperioder}
\usage{
tabAntOpphEnhTid(
  RegData,
  datoTil = Sys.Date(),
  enhetsNivaa = "ShNavn",
  tidsEnhet = "Mnd",
  antTidsenh = 6
)
}
\arguments{
\item{RegData}{dataramme}

\item{datoTil}{siste registrering som vises}

\item{enhetsNivaa}{Aggregeringsnivå RHF, HF, ShNavn (sykehus, standard)}

\item{tidsEnhet}{- Mnd, Kvartal, Aar}

\item{antTidsenh}{antall måneder som skal vises}
}
\description{
tabAntOpphEnhTid antall opphold siste X (antMnd) mnd
RegData må inneholde ikke-aggregerte data, dvs. data på oppholdsnivå
Summerer antall opphold for hele tidsperioder
}
