% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{tabAntPersOpph}
\alias{tabAntPersOpph}
\title{Title}
\usage{
tabAntPersOpph(RegData, datoFra, datoTil = Sys.Date(), enhetsNivaa)
}
\arguments{
\item{RegData}{dataramme}

\item{datoFra}{startdato}

\item{datoTil}{sluttdato}

\item{enhetsNivaa}{'HF' eller 'RHF'}
}
\description{
Title
}
