% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{TabTidEnhet}
\alias{TabTidEnhet}
\title{Antall tilfeller for valgt tidsenhet og enhetsnivå. Filtreringer kan også gjøres.}
\usage{
TabTidEnhet(
  RegData,
  tidsenhet = "dag",
  erMann = 9,
  enhetsNivaa = "RHF",
  bekr = 9,
  skjemastatus = 9,
  dodSh = 9,
  valgtEnhet = "Alle",
  velgAvd = 0
)
}
\arguments{
\item{RegData}{dataramme med preprossesserte data}

\item{tidsenhet}{'dag' (standard), 'uke', 'maaned'}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{enhetsNivaa}{'RHF', 'HF', 'ShNavn'}

\item{bekr}{1-bekreftet, 0-mistenkt tilfelle}

\item{skjemastatus}{status på registreringa 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{dodSh}{død på sykehus 0-nei, 1-ja}
}
\value{

}
\description{
Antall tilfeller for valgt tidsenhet og enhetsnivå. Filtreringer kan også gjøres.
}
