% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabellerK2n.R
\name{tr_summarize_output}
\alias{tr_summarize_output}
\title{Transponer output fra tidyr::summarize}
\usage{
tr_summarize_output(x, grvarnavn = "")
}
\arguments{
\item{x}{En dataramme med output fra en gruppert summarise}

\item{grvarnavn}{Navnet som skal gis til første kolonne i output}
}
\value{
tr_frame Den transponerte av inputen
}
\description{
Denne funksjonen tar som input resultatet av tidyr::summarize og returnerer dens
transponerte uten at formatene endres.
}
