% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{AlderTab}
\alias{AlderTab}
\title{Aldersfordeling, tabell}
\usage{
AlderTab(
  RegData,
  valgtEnhet = "Alle",
  enhetsNivaa = "Alle",
  skjemastatus = 9,
  dodSh = 9,
  erMann = 9
)
}
\arguments{
\item{RegData}{datatabell, beredskapsdata}

\item{enhetsNivaa}{styres av tilgangsnivå 'Alle', 'RHF', 'HF'}

\item{skjemastatus}{status på registreringa 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{dodSh}{død på sykehus 0-nei, 1-ja}

\item{erMann}{kjønn, 0-kvinne, 1-mann}
}
\value{

}
\description{
Aldersfordeling, tabell
}
