% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabellerK2n.R
\name{antallTidBelegg}
\alias{antallTidBelegg}
\title{Belegg per tidsenhet  og HF. Filtreringer kan også gjøres.
Detaljerinsnivå er styrt av tilgangsnivå}
\usage{
antallTidBelegg(
  RegData,
  tidsenhet = "dag",
  erMann = 9,
  tilgangsNivaa = "SC",
  datoFra = 0,
  datoTil = Sys.Date(),
  skjemastatusInn = 9,
  aarsakInn = 9,
  valgtEnhet = "Alle",
  reshID = 0
)
}
\arguments{
\item{RegData}{dataramme med preprossesserte data}

\item{tidsenhet}{'dag' (standard), 'uke', 'maaned', 'aar'}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{tilgangsNivaa}{SC, LC og LU bestemmer hvilket enhetsNivaa
('RHF', 'HF', 'ShNavn') resultatene skal vises for}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{skjemastatusInn}{status på inklusjonsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{aarsakInn}{covid-19 som hovedårsak til innleggelse 1-ja, alle opph, 2-ja, minst siste opphold,
3-ja, minst ett opph, 4-nei, ingen opph, 9-ukj}

\item{valgtEnhet}{NULL for SC-bruker, ellers eget RHF/HF}
}
\description{
Belegg per tidsenhet  og HF. Filtreringer kan også gjøres.
Detaljerinsnivå er styrt av tilgangsnivå
}
