% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{antallTidEnhTab}
\alias{antallTidEnhTab}
\title{Antall tilfeller for valgt tidsenhet og enhetsnivå. Filtreringer kan også gjøres.
Detaljerinsnivå er styrt av tilgangsnivå}
\usage{
antallTidEnhTab(
  RegData,
  tidsenhet = "dag",
  erMann = 9,
  datoFra = 0,
  datoTil = Sys.Date(),
  tilgangsNivaa = "SC",
  valgtEnhet = "Alle",
  skjemastatusInn = 9,
  aarsakInn = 9,
  dodSh = 9
)
}
\arguments{
\item{RegData}{dataramme med preprossesserte data}

\item{tidsenhet}{'dag' (standard), 'uke', 'maaned'}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{tilgangsNivaa}{SC, LC og LU bestemmer hvilket enhetsNivaa}

\item{valgtEnhet}{NULL for SC-bruker, ellers eget RHF/HF}

\item{skjemastatusInn}{status på inklusjonsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{aarsakInn}{covid-19 som hovedårsak til innleggelse 1-ja, alle opph, 2-ja, minst siste opphold,
3-ja, minst ett opph, 5-nei, ingen opph, 9-ukj}

\item{dodSh}{død på sykehus 1-'nei', 2-'ja', 3-'ja og nei' (=ferdigstilte?)}

\item{HF}{benytte kortnavn for HF 0-nei, 1-ja
('RHF', 'HF', 'ShNavn') resultatene skal vises for}
}
\value{

}
\description{
Antall tilfeller for valgt tidsenhet og enhetsnivå. Filtreringer kan også gjøres.
Detaljerinsnivå er styrt av tilgangsnivå
}
