% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaHjelpefunk.R
\name{erInneliggende}
\alias{erInneliggende}
\title{Funksjon som avgjør om en pasient er inneliggende på aktuell dato}
\usage{
erInneliggende(datoer, regdata)
}
\arguments{
\item{datoer}{datoer som inneligging skal avgjøres for}

\item{regdata}{Dataramme som inneholder InnDato og Utdato per pasient}
}
\value{

}
\description{
Returnerer TRUE for datoer pasienten er inneliggende
}
