% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{FerdigeRegInnTab}
\alias{FerdigeRegInnTab}
\title{Ferdigstilte registreringer}
\usage{
FerdigeRegInnTab(
  RegData,
  valgtEnhet = "Alle",
  enhetsNivaa = "RHF",
  erMann = 9,
  dodSh = 9
)
}
\arguments{
\item{RegData}{beredskapsskjema}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{dodSh}{død på sykehus 0-nei, 1-ja}
}
\value{

}
\description{
Ferdigstilte registreringer
}
