% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunksjonerDatafilerFHI.R
\name{hentPandemiDataFHI}
\alias{hentPandemiDataFHI}
\title{Henter data og velger variabler for overføring til FHI}
\usage{
hentPandemiDataFHI(personIDvar = "PersonIdBC19Hash")
}
\value{
Pandemidata tilrettelagt for FHI
}
\description{
Henter data og velger variabler for overføring til FHI
}
