% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaDataSQL.R
\name{KoronaDataSQL}
\alias{KoronaDataSQL}
\title{Henter data fra database for koronaregisteringer.}
\usage{
KoronaDataSQL(
  datoFra = "2020-03-01",
  datoTil = Sys.Date(),
  skjema = 1,
  koble = 1
)
}
\arguments{
\item{skjema}{1-innleggelse, 2-utskriving}

\item{koble}{0-kun valgt skjema, 1 koble inn og ut-skjema}
}
\value{
Henter dataramma RegData for Pandemiregisteret
}
\description{
Henter data fra database for koronaregisteringer.
}
