% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaUtvalg.R
\name{KoronaUtvalg}
\alias{KoronaUtvalg}
\title{Gjøre utvalg i koronaskjema}
\usage{
KoronaUtvalg(
  RegData,
  datoFra = 0,
  datoTil = 0,
  erMann = 9,
  minald = 0,
  maxald = 110,
  skjemastatusInn = 9,
  skjemastatusUt = 9,
  dodSh = 9,
  aarsakInn = 9,
  enhetsNivaa = "RHF",
  valgtEnhet = "Alle",
  enhetsUtvalg = 2,
  beredPas = 9
)
}
\arguments{
\item{RegData}{data, beredskapsskjema}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{minald}{kjønn, 0-kvinne, 1-mann}

\item{skjemastatusInn}{status på inklusjonsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{skjemastatusUt}{status på utskrivingsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{dodSh}{død på sykehus 1-'nei', 2-'ja', 3-'ja og nei' (=ferdigstilte?)}

\item{aarsakInn}{covid-19 som hovedårsak til innleggelse 1-ja, alle opph, 2-ja, minst siste opphold,
3-ja, minst ett opph, 5-nei, ingen opph, 9-ukj}

\item{enhetsNivaa}{organisatorisk nivå, dvs. filtreringsnivå for egen enhet.
RHF (SC og LC-brukere), HF (LU-brukere)}

\item{valgtEnhet}{- gis ut fra resh}

\item{enhetsUtvalg}{0: hele landet,
1: egen enhet mot resten av landet,
2: egen enhet (NB: returnerer hele landet i ind$Rest)}
}
\description{
Returnerer filtrert dataramme og utvalgstekst.
NB: RegData filtreres IKKE på enhet. Velg enhetsutvalg 1 el 2 for
å få ind$Hoved for "egen enhet"
}
\details{
Argumentet \emph{enhetsUtvalg} har følgende valgmuligheter:
   \itemize{
    \item 0: Hele landet
    \item 1: Egen enhet mot resten av landet (Standard)
    \item 2: Egen enhet (kan være både RHF eller HF)
   	}
}
