% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatafilerFHI.R
\name{lagDatafilerTilFHI}
\alias{lagDatafilerTilFHI}
\title{Henter data og tilrettelegger filer for overføring til FHI}
\usage{
lagDatafilerTilFHI()
}
\value{

}
\description{
Henter data og tilrettelegger filer for overføring til FHI
}
