% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluensadataFHI.R
\name{lagInfluDataFHI}
\alias{lagInfluDataFHI}
\title{Henter data og velger variabler for overføring til FHI}
\usage{
lagInfluDataFHI()
}
\description{
Henter data og velger variabler for overføring til FHI
}
