% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{RisikoInnTab}
\alias{RisikoInnTab}
\title{Tabell med oversikt over tilstander som medfører økt risiko ved Coronasmitte}
\usage{
RisikoInnTab(
  RegData,
  datoFra = "2020-03-01",
  datoTil = Sys.Date(),
  erMann = "",
  skjemastatusInn = 9,
  dodSh = 9,
  aarsakInn = 9,
  valgtEnhet = "Alle",
  enhetsNivaa = "RHF",
  minald = 0,
  maxald = 110
)
}
\arguments{
\item{RegData}{data}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{skjemastatusInn}{status på inklusjonsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{dodSh}{død på sykehus 1-'nei', 2-'ja', 3-'ja og nei' (=ferdigstilte?)}

\item{aarsakInn}{covid-19 som hovedårsak til innleggelse 1-ja, alle opph, 2-ja, minst siste opphold,
3-ja, minst ett opph, 5-nei, ingen opph, 9-ukj}

\item{valgtEnhet}{- gis ut fra resh}

\item{enhetsNivaa}{organisatorisk nivå, dvs. filtreringsnivå for egen enhet.
RHF (SC og LC-brukere), HF (LU-brukere)}

\item{minald}{kjønn, 0-kvinne, 1-mann}
}
\value{

}
\description{
Tabell med oversikt over tilstander som medfører økt risiko ved Coronasmitte
}
