% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{RisikoInnTab}
\alias{RisikoInnTab}
\title{Tabell med oversikt over tilstander som medfører økt risiko ved Coronasmitte}
\usage{
RisikoInnTab(
  RegData,
  datoTil = Sys.Date(),
  reshID = 0,
  erMann = "",
  skjemastatusInn = 9,
  dodSh = 9,
  aarsakInn = 9,
  valgtEnhet = "Alle",
  enhetsNivaa = "RHF",
  minald = 0,
  maxald = 110
)
}
\arguments{
\item{RegData}{data}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{reshID}{reshID fra innlogging}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{skjemastatusInn}{status på inklusjonsskjema 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{dodSh}{død på sykehus 0-nei, 1-ja}

\item{aarsakInn}{covid-19 som hovedårsak til innleggelse}
}
\value{
Risikofaktorer (av alle)' = AntAndel(RegData$KjentRisikofaktor==1, dim(RegData)[1])
}
\description{
Tabell med oversikt over tilstander som medfører økt risiko ved Coronasmitte
}
