% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaHjelpefunk.R
\name{sendDataFilerFHI}
\alias{sendDataFilerFHI}
\title{Funksjon som henter filer som skal sendes til FHI. To filer fra intensivopphold
og to filer fra sykehusopphold. Dvs. Ei fil for hvert opphold og ei aggregert til
person, for hvert register}
\usage{
sendDataFilerFHI(
  rnwFil,
  brukernavn = "lluring",
  Rpakke = "korona",
  rolle = "SC"
)
}
\arguments{
\item{rnwFil}{Navn på fila som skal kjøres.}

\item{Rpakke}{hvilken R-pakke fila som lager rapporten ligger i}

\item{parametre}{Liste med valgfrie parametre, avhengig av type rapport}
}
\value{
Full path of file produced
}
\description{
Funksjon som henter filer som skal sendes til FHI. To filer fra intensivopphold
og to filer fra sykehusopphold. Dvs. Ei fil for hvert opphold og ei aggregert til
person, for hvert register
}
