% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{statusNaaTab}
\alias{statusNaaTab}
\title{Antall som er  i ECMO/respirator}
\usage{
statusNaaTab(
  RegData,
  valgtEnhet = "Alle",
  enhetsNivaa = "RHF",
  erMann = 9,
  bekr = 9
)
}
\arguments{
\item{RegData}{beredskapsskjema}
}
\value{

}
\description{
Antall som er  i ECMO/respirator
}
