% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{statusNaaTab}
\alias{statusNaaTab}
\title{Status nå}
\usage{
statusNaaTab(
  RegData,
  valgtEnhet = "Alle",
  enhetsNivaa = "RHF",
  aarsakInn = 9,
  erMann = 9
)
}
\arguments{
\item{RegData}{pandemiskjema}
}
\description{
Status nå
}
