% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoronaTabeller.R
\name{tabAntPersOpph}
\alias{tabAntPersOpph}
\title{Antall personer, smitteforløp, forløp i samme tabell per HF}
\usage{
tabAntPersOpph(
  RegData,
  datoFra,
  datoTil = Sys.Date(),
  enhetsNivaa,
  covidInn = 0
)
}
\arguments{
\item{RegData}{dataramme}

\item{datoFra}{startdato}

\item{datoTil}{sluttdato}

\item{enhetsNivaa}{'HF' eller 'RHF'}

\item{covidInn}{covid hovedårsak til innleggelse? 0-alle reg., 1-ja, 2-nei}
}
\description{
Antall personer, smitteforløp, forløp i samme tabell per HF
}
